requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.2
sqlalchemy>=2.0.25
PyMuPDF>=1.23.0
pandas>=2.0.0